% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bystrataCoreFuns.R
\name{adaptivewtHR}
\alias{adaptivewtHR}
\title{Estimate amalgamated treatment effect}
\usage{
adaptivewtHR(weights, betas, vars, cilevel)
}
\arguments{
\item{weights}{Original, non-adaptive weights (e.g., sample size weights)}

\item{betas}{Vector of estimated treatment effect within each strata}

\item{vars}{Vector of estimated variances of betas}

\item{cilevel}{Significance level for adaptive weighted confidence intervals
(i.e., (1-cilevel)x100\% CIs)}
}
\value{
\itemize{
    \item adaptivewtRes: table of amalgamated log-hazard estimate, variance,
    cilevelx100\% ci,and p-value using the adaptive weight
    \item adaptivewts: adaptive weights used
    \item adaptivecorr: rank correlation between betas and vars used to determine
    where original (i.e., combining estimates) or new weights (i.e., combining
    test statistics) are used
}
}
\description{
Estimates overall treatment effect amalgamated from adaptively weighted
strata-level treatment effects
}
