% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bystrataCoreFuns.R
\name{eventsbystrata}
\alias{eventsbystrata}
\title{Summarize events per strata x arm}
\usage{
eventsbystrata(treeStrata, arm, status, family = "cox")
}
\arguments{
\item{treeStrata}{Vector of subject-level strata membership}

\item{arm}{Treatment indicator, 1 = test treatment, 0 = control}

\item{status}{For family="cox", the censoring status (1 = event,
0 = censored); for family="binomial" the case/control status}

\item{family}{Trait family, current options: "cox" or "binomial"}
}
\value{
Table summarizing number of events, subjects, and percent censoring
(for family = "cox") or percent controls (for family = "binomial")
within each strata and treatment arm
}
\description{
Summarizes number of events, subjects, and percent censoring within each
strata and treatment arm from formed risk strata
}
