% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5STARcorefun.R
\name{filter5STAR}
\alias{filter5STAR}
\title{Filter covariates for 5STAR}
\usage{
filter5STAR(yy, X, family = "cox", cilevel = 0.05, plot = FALSE,
  verbose = 0, filter.hyper = filter_control(method = "ENET",
  lambdatype = "min", mixparm = NULL, ...))
}
\arguments{
\item{yy}{Response - either Surv() object for time to event data or 1
column matrix of case/control status for binary data}

\item{X}{Data frame of all possible stratification covariates}

\item{family}{Trait family, current options: "cox", "binomial", or "gaussian"}

\item{cilevel}{Optional significance level for RF VIMP confidence intervals}

\item{plot}{Whether to make include VIMP CI plots (ignored if
method = "ENET")}

\item{verbose}{Numeric variable indicating amount of information to print
to the terminal (0 = nothing, 1 = notes only, 2 = notes and intermediate output)}

\item{filter_control}{List of contorl parameters for filtering step, @seealso
\code{\link{filter_control}}; key agruments include: \itemize{
\item method: filtering method; current options are: "ENET" (Elastic Net),
"RF" (Random Forest with default or input parameters, using delete-d
jackknife confidence intervals for VIMP for variable selection), "RFbest"
(Random Survival Forest, performing tuning for mtry and nodesize, and using
 double bootstrap confidence intervals for VIMP for variable selection;
 more accurate but slower than RF option)
 \item mixparm: Optional elastic net mixing parameter alpha or grid of alpha
values to search over. If nothing is entered, will search for best value
between 0.05 and 0.95. Ignored when method is "RF" or "RFbest"
 \item lambdatype: Optional elastic net parameter; whether to use lambda that
minimizes cross validation error (lambdatype="min") or the largest lambda
that gives error within 1 standard error of minimum error (lambdatype="1se",
default). Ignored when method is "RF" or "RFbest"
 \item ... : Optional arguments for glmnet or rfsrc
}}
}
\value{
cov2keep: List of all covariates to keep after filtering step, selected by
elastic net or random forest

For method=ENET, additionally outputs \itemize{
   \item cvout: containing fraction of null deviance explained, mean cross validation
error, and optimal lambda and alpha value (see cv.glmnet and glmnet for
more details)
   \item beta: the coefficients of glmnet fit given tuned choice of alpha and lambda
}

For method=RF or RFbest, additionally outputs:
\itemize{
   \item varselectmat:  matrix of VIMP confidence intervals, p-values, and selection
decision for each variable
   \item VIMPplot: default variable importance CI plot, output if plot = TRUE
   \item varselect2plot: variable selection information from subsample.rfsrc for
making customizable VIMP CI plots
   \item forest: rfsrc object
}
}
\description{
Forms dummy variable matrix for factors, and fits an elastic net or random
forest model to determine which covariates to keep for building trees in 5-STAR
}
