% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bystrataCoreFuns.R
\name{glmbystrata}
\alias{glmbystrata}
\title{Fit Linear or Logistic Regression GLM Model Within Strata}
\usage{
glmbystrata(yy, arm, family, treeStrata, termNodes = NULL,
  cilevel = 0.05, verbose = 0, alternative = "two.sided")
}
\arguments{
\item{yy}{Continuous response or case control status response vector}

\item{arm}{Treatment indicator, 1 = test treatment, 0 = control}

\item{family}{Family for glm; options: "binomial" or "gaussian"}

\item{treeStrata}{Vector of strata membership, as defined by ctree strata}

\item{termNodes}{Vector of names for each tree strata formed (e.g., number of
strata or covariate definition)}

\item{cilevel}{Confidence level alpha for overall result and confidence
intervals}

\item{verbose}{Numeric variable indicating amount of information to print
to the terminal (0 = nothing, 1 = notes only, 2 = notes and intermediate output)}
}
\value{
\itemize{
    \item fitsummary: summary of GLM fits within each strata
    \item table: Summary of amalgamated regression coefficient or odds ratio
     estimate, variance, p-value, and cilevelx100\% confidence interval
     assuming sample size weights
    \item weights: Sample size weights used to construct estimate
}
}
\description{
Fits a linear or logistic regression GLM model within each formed strata,
providing estimate of regression coefficient or log odds ratio (EXPERIMENTAL!)
}
