% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bystrataCoreFuns.R
\name{mdbystrata}
\alias{mdbystrata}
\title{Estimate mean difference Within Strata}
\usage{
mdbystrata(yy, arm, treeStrata, termNodes = NULL, cilevel = 0.05,
  verbose = 0, alternative = "two.sided")
}
\arguments{
\item{yy}{Continuous response vector}

\item{arm}{Treatment indicator, 1 = test treatment, 0 = control}

\item{treeStrata}{Vector of strata membership, as defined by ctree strata}

\item{termNodes}{Vector of names for each tree strata formed (e.g., number of
strata or covariate definition)}

\item{cilevel}{Confidence level alpha for overall result and confidence
intervals}

\item{verbose}{Numeric variable indicating amount of information to print
to the terminal (0 = nothing, 1 = notes only, 2 = notes and intermediate output)}

\item{treetype}{String, whether trees input are "preliminary" or "final"}

\item{plot}{Logical, whether to create risk difference forest plots}
}
\value{
\itemize{
    \item fitsummary: summary of mean difference within each strata
    \item table: Summary of amalgamated mean difference estimate, variance,
    p-value, and cilevelx100\% confidence interval assuming sample size weights
    \item weights: Sample size weights used to construct estimate
}
}
\description{
Estimates difference in means and significance within each formed strata
}
