% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bystrataCoreFuns.R
\name{minPadapHR}
\alias{minPadapHR}
\title{Estimate amalgamated treatment effect using minP approach}
\usage{
minPadapHR(sf, betas, vars, cilevel, alternative = "less",
  vartype = "alt")
}
\arguments{
\item{sf}{A \code{\link[survival]{survfit}}
object containing information on number of subjects,
number of events, number at risk, etc. for each strata, pooled across
treatment assignment}

\item{betas}{Vector of estimated (log) treatment effect within each strata}

\item{vars}{Vector of estimated variances of betas}

\item{cilevel}{Significance level for adaptive weighted confidence intervals
(i.e., (1-cilevel)x100\% CIs for two-tailed tests, and (1-2*cilevel)x100\% CIs
for 1-tailed tests) (default = 0.025)}

\item{alternative}{For tests, whether alternative hypothesis is "less",
"greater", or "two.sided" (default = "less")}

\item{vartype}{Whether stratum-level variances used to calculate the correlation
between sample size and ni/sqrt(Vi) test statistics should be calculated
under the null hypothesis ("null") or estimated via the fitted model under the
alternative hypothesis ("alt", default). "null" variance may only be used when
measure == "HR"}
}
\value{
\itemize{
    \item adaptivewtRes: table of amalgamated (log) treatment effect estimate,
     variance, ci,and p-value using the optimal weight, as well as
     probability the treatment effect is in the desired direction
    \item adaptivewts: weights used (e.g., ni or ni/sqrt(Vi))
    \item calpha: adjusted critical value for test accounting for selecting
    the optimal weighting scheme
    \item singlewtres: vector of Z-score test statistics and corresponding
    p-values for sample size weights, ni/sqrt(Vi) weights, and inverse variance
    weights (1/Vi; output for knowledge though not used in calculations)
}
}
\description{
Estimates overall treatment effect amalgamated from strata-level treatment
effect estimates, using optimal weighting (i.e., using the the weighting scheme
that gives the best test statistics/minimum p-value, between ni and ni/sqrt(Vi)
weights), paying a minor penalty for taking the best of two correlated test
statistics
}
