% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5STARcorefun.R
\name{prep5STAR}
\alias{prep5STAR}
\title{Prepares data for 5-STAR algorithm}
\usage{
prep5STAR(yy, X, family = "cox", missthreshold = c(0.1, 0.2),
  verbose = 0)
}
\arguments{
\item{yy}{Trait/response (currently must be either continous, binary, or
a Surv() object summarizing follow-up time for right-censored data and status
 indicator where 1=dead, 0=censored). Ignored for family != "cox"}

\item{X}{Data frame of all possible stratification covariates}

\item{family}{Trait family, current options: "cox", "binomial", or "gaussian"}

\item{missthreshold}{Vector of lower and upper bound of acceptable missingness
levels for each covariate, such that covariates with less than the first
missthreshold value will be passed to step 2 (filtering step), those with
missingess greater than the second missthreshold value will be removed from
analysis, and those with missingness between these two values will be included
only if they are significantly correlated with the outcome scores
(e.g., logrank scores). If a scalar is entered, covariates with less than
that amount of missingness will be included and those with greater will be
removed. For family other than "cox", only the first value is used.}

\item{verbose}{Numeric variable indicating amount of information to print
to the terminal (0 = nothing, 1 = notes only, 2 = notes and intermediate output)}
}
\value{
\itemize{
\item X: cleaned covariate matrix
\item input covs: names of all covariates which will be passed into filtering
step of 5-STAR algorithm
}
}
\description{
Cleans covariate matrix, removing covariates with too much missingess,
and coverts all character covariates into factors
}
