% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5STARcorefun.R
\name{run5STAR}
\alias{run5STAR}
\title{Run 5-STAR Algorithm}
\usage{
run5STAR(yy, arm, X, family = "cox", measure = "OR", cilevel = 0.05,
  missthreshold = c(0.1, 0.2), verbose = 0, plot = TRUE,
  timeunit = NULL, filter.hyper = filter_control(),
  tree.hyper = tree_control())
}
\arguments{
\item{yy}{Trait/response (currently must be either continous, binary, or
a Surv() object summarizing follow-up time for right-censored data and status
 indicator where 1=dead, 0=censored)}

\item{arm}{Binary treatment indicator, 1 = test treatment, 0 = control}

\item{X}{Data frame of all possible stratification covariates}

\item{family}{Trait family, current options: "cox", "binomial", and
"gaussian" (note: binomial and gaussian are still experimental!)}

\item{measure}{Response of interest for binary traits; current options
are: "RD" (rate difference using Miettinen and Nurminen 1985 method), or
"OR" (odds ratio, using GLM model fit); ignored for family!="binomial"}

\item{cilevel}{Confidence level alpha for overall result and confidence
intervals}

\item{verbose}{Numeric variable indicating amount of information to print
to the terminal (0 = nothing, 1 = notes only, 2 = notes and intermediate output)}

\item{plot}{Logical, whether to return summary plots (within strata and
pooled KM plots, forest plots summarizing strata-level and overall results)}

\item{timeunit}{Optional parameter for plots defining scale of time to event
data (e.g., "Months", "Years"); ignored for family!="cox"}

\item{filter.hyper}{A list of control parameters for filtering step,
(see \code{\link{filter_control}})}

\item{tree.hyper}{A list of control parameters for strata building step
(see \code{\link{tree_control}})}

\item{missthrehold}{Optional parameter specifying proportion of missingness
allowed for each covariate. See \code{\link{prep5STAR}} for more details.}
}
\value{
\itemize{
    \item inputcovs: vector containing names of all covariates input into
    the algorithm (after removing those with too much missingness, as defined
    by missthreshold)
    \item filteredcovs: vector containing all covariates left after
    filtering step
    \item filterdetails: list containing additional output from filter step;
    for method=ENET, this includes: \itemize{
   \item cvout: containing fraction of null deviance explained, mean cross validation
error, and optimal lambda and alpha value (see cv.glmnet and glmnet for
more details)
   \item beta: the coefficients of glmnet fit given tuned choice of alpha and lambda
} For method=RF or RFbest, this includes:
\itemize{
   \item varselectmat: matrix of VIMP confidence intervals, p-values, and selection
decision for each variable
   \item VIMPplot: default variable importance CI plot, output if plot = TRUE
   \item varselect2plot: variable selection information from subsample.rfsrc for
making customizable VIMP CI plots
   \item forest: rfsrc object
}
    \item prelimtree: preliminary tree (3A) built by ctree, in terms
     of covariates
    \item prelimstratadefn: vector of rules defining the preliminary strata
    (step 3A), as defined by covariates chosen in initial ctree run, ordered from
    highest to lowest risk
    \item prelimstrataids: vector of preliminary (3A) strata membership for
    each subject (where 1 indicates highest risk, and the largest number
    indicates lowest risk, i.e., matching order of prelimstratadefn)
    \item prelimstratasurvfits: a \code{\link[survival]{survfit}} object,
    summarizing survival curves in each preliminary (3A) strata, pooled by
    treatment assignment; returned when family="cox"
    \item prelimbystratasummary: matrix summarizing  cox/glm fits within
    each preliminary (3A) strata, including estimated (log) hazard ratio
    (or OR, RD, etc. when family != "cox"), corresponding (1-cilevel)x100\%
    CI, Gramsch and Therneau test of non-PH p-value, and sample size,
    inverse variance, and adaptive weights
    \item prelimadaptivecorr: rank correlation between estimated coefficients
    beta and corresponding variances from each preliminary (3A) strata
    \item prelimres5star: Summary of amalgamated hazard ratio (family="cox") or
    odds ratio (family="binomial") estimate, variance,
    p-value, and (1-cilevel)x100\% confidence interval for adaptive weights
    from preliminary strata (3A)
    \item finaltree: final, pruned tree (3B) built by ctree, in terms of
     preliminary strata
    \item stratadefn: vector of rules defining the final strata
    (step 3B), as defined by covariates translated from pruning ctree run,
    ordered from highest to lowest risk
    \item strataids:  vector of final (3B) strata membership for
    each subject (where 1 indicates highest risk, and the largest number
    indicates lowest risk, i.e., matching order of stratadefn)
    \item stratasurvfits: a \code{\link[survival]{survfit}} object,
    summarizing survival curves in each final (3B) strata, pooled by
    treatment assignment; returned when family="cox"
    \item bystratasummary: matrix summarizing  cox/glm fits within
    each final (3B) strata, including estimated (log) hazard ratio (or OR,
    RD, etc. when family != "cox"), corresponding (1-cilevel)x100\% CI,
    Gramsch and Therneau test of non-PH p-value, and sample size,
    inverse variance, and adaptive weights
    \item adaptivecorr: rank correlation between estimated coefficients
    beta and corresponding variances from each final (3B) strata; used
    to determine if adaptive weights will combine betas or test statistics
    \item res5star: Summary of amalgamated hazard ratio (family="cox") or
    odds ratio (family="binomial") estimate, variance,
    p-value, and cilevelx100% confidence interval for adaptive weights
    \item prelimbetweenstrataKM: Kaplan-Meier survival curve plots from pooled
     treatment assignment data from each preliminary (3A) strata,
     returned if plot = TRUE and family = "cox"
    \item prelimbystrataKM: Kaplan-Meier survival curve plots within each
    preliminary (3A) strata, returned if plot = TRUE and family = "cox"
    \item prelimforestplot: a forest plot summarizing estimated coefficient and
    confidence interval for each final (3B) strata, as well as amalgamated
    result
    \item betweenstrataKM: Kaplan-Meier survival curve plots from pooled
     treatment assignment data from each final (3B) strata,
     returned if plot = TRUE and family = "cox"
    \item bystrataKM: Kaplan-Meier survival curve plots within each final
    (3B) strata, returned if plot = TRUE and family = "cox"
    \item forestplot: a forest plot summarizing estimated coefficient and
    confidence interval for each final (3B) strata, as well as amalgamated
    result
}
}
\description{
Runs 5-STAR algorithm for forming homogeneous risk strata and calculating
an overall treatment effect estimate amalgamated from the formed strata
}
\details{
Filtering step is performed on complete case basis (e.g., removing
all individuals with any missing covariate data)
}
\seealso{
\code{\link{filter_control}}

\code{\link{tree_control}}
}
