% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bystrataCoreFuns.R
\name{strataforestplot}
\alias{strataforestplot}
\title{Make forest plot to summarize strata fits}
\usage{
strataforestplot(MRSSmat, MRSSres, stratafits, family, measure,
  treetype = "final", labelNames = NULL, cilevel, ...)
}
\arguments{
\item{MRSSmat}{matrix summary of model (e.g., cox, glm) fits within each strata
with required columns "exp.bhat", "exp.ci.lower.", and "exp.ci.upper." (e.g.,
HR/OR and corresponding confidence interval) and "weight.adap"
for family = "cox" or "binomial",
measure="OR, or columns "ratediff", "ci.lower", "ci.upper" for
family = "binomial", measure="RD", or columns "bhat", "ci.lower",
"ci.upper" for family = "gaussian"}

\item{MRSSres}{row matrix with amalgamated results summary, including estimate
and corresponding confidence intervals}

\item{stratafits}{survfit object with number of subjects, median survival
times, etc. within each strata}

\item{family}{Trait family, current options: "cox", "binomial, "gaussian"}

\item{measure}{Response of interest for binary traits; current options
are: "RD" (rate difference using Miettinen and Nurminen 1985 method), or
"OR" (odds ratio, using GLM model fit); ignored for family!="binomial"}

\item{treetype}{String, whether trees input are "preliminary" or "final"}

\item{labelNames}{Vector of names for each tree strata formed (e.g., number of
strata or covariate definition)}

\item{cilevel}{Confidence level alpha for overall result and confidence
intervals}

\item{...}{Optional additional arguments passed into the forestplot function
(@seealso \code{\link[forestplot]{forestplot}})}
}
\value{
fplot An annotated forest plot showing by-stratum estimates and
 confidence intervals, as well as amalgamated result
}
\description{
Make a forest plot to summarize model fits within each strata
}
