% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bystrataCoreFuns.R
\name{coxbystrata}
\alias{coxbystrata}
\title{Fit CoxPH Model Within Strata}
\usage{
coxbystrata(time, status, arm, treeStrata, termNodes = NULL,
  treetype = "final", cilevel = 0.05, verbose = 0, plot = TRUE,
  timeunit = NULL)
}
\arguments{
\item{time}{Follow-up time for right-censored data}

\item{status}{Status indicator, 1=event, 0=censored}

\item{arm}{Treatment indicator, 1 = test treatment, 0 = control}

\item{treeStrata}{Vector of strata membership, as defined by ctree strata}

\item{termNodes}{Vector of names for each tree strata formed (e.g., number of
strata or covariate definition)}

\item{treetype}{String, whether trees input are "preliminary" (e.g.,
from step 3A) or "final" (e.g., from step 3B)}

\item{cilevel}{Confidence level alpha for overall result and confidence
intervals (i.e., produces (1-cilevel)x100\% CIs)}

\item{verbose}{Numeric variable indicating amount of information to print
to the terminal (0 = nothing, 1 = notes only, 2 = notes and intermediate output)}

\item{plot}{Logical, whether to create within strata and pooled between
strata KM plots}

\item{timeunit}{Optional argument, time unit for survival data
(e.g., Months, Years,..); currently only used for plots and ignored if
plot is FALSE}
}
\value{
\itemize{
    \item fitsummary: summary of cox fits within each strata, along with
    Grambsch-Therneau (GT) tests for PH within each strata
    \item stratafit: list of full coxph fits within each strata
    \item table: Summary of amalgamated hazard ratio estimate, variance,
    p-value, and cilevelx100\% confidence interval assuming sample size weights
    \item weights: Sample size weights used to construct estimate
    \item bystrataKM: Kaplan-Meier survival curve plots within each strata,
    returned if plot = TRUE
    \item betweenstrataKM: Kaplan-Meier survival curve plots from pooled
     treatment assignment data from each strata, returned if plot = TRUE
}
}
\description{
Fits a Cox proportional hazards model within each formed strata,
providing estimate of log hazard ratio and test of PH within each strata
}
